Initial Prufer's Decoding Algorithm Script
This is the starting script for the Prufer's Decoding Algorithm.

View the README file to understand how the algorithm works and how to run the script.

One thing that may not be clear is the if statements with degreesLeft > 2.

I'll explain with an example. Suppose the sequence <6,2,4,6>

The way the script is set up to take in the sequence and split the sequence to have each vertex in the sequence into its own sublist and have one extra sublist for the last 2 remaining indexes in the degrees list.
-Sequence list in script: [ [6],[2],[4],[6],[]]
-Degree list: [1,2,1,2,1,3] (Index = vertex - 1 and degrees per vertex = number of appearances in sequence + 1)

The algorithm will pair up vertices to the vertices in the sublists. In the end, after all the pairings have been made with the vertices that were in the sublists, only the last sublist remains emptry for the two vertices whose index has the value of 1.
-Sequence list in script: [ [1,6],[3,2],[2,4],[4,6],[]]
-Degree list: [0,0,0,0,1,1]

Since we want the last 2 vertices to be recorded in the last sublist, the if statements are required so that the algorithm stays at the last sublist to store the two remaining vertices, unlike the previous iterations where we only stored 1 vertices per sublist. That is the purpose of a few of the if statements, to take into account of that last sublist to store the final edge.